+++ lib/workitcop.rb
@@ -3,6 +3,7 @@
 require_relative "workitcop/inject"
 require_relative "workitcop/version"
 
+require_relative "rubocop/cop/workit/action_args"
 require_relative "rubocop/cop/workit/restrict_on_send"
 
 module Workitcop