+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add new `Workit/ActionArgs` cop. ([@ydah][])
+
 ## [0.1.0] - 2022-07-04
 
 - Initial release